client/python/requirements.txt
numpy==1.19.5
protobuf==3.20.1
websocket-client==1.4.2
pillow 
opencv-python
